Task2/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
requests==2.31.0
httpx==0.26.0

# LLM Integration
openai==1.10.0

# File Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
tenacity==8.2.3

# Deployment
gunicorn==21.2.0